AddControlData {OCTOPUS}	R Documentation

AddControlData

Description

Used to include control data.

Usage

AddControlData(cISAData, cSourceData, nISA)

Arguments

cISAData
The ISA object, the class( cISAData ) will determine which version of AddControlData is utilized.

cSourceData
The object with the patient data to be used as the patient database for pulling control data from. THis object must have the following elements: 1) cSourceData$vIND 2) cSourceData$vOut 3) cSourceData$vTime 4) cSourceData$vTrt

nISA
An integer for the ISA number. As the data is added to the control in may be necessary to know what ISA the data came from.

[Package OCTOPUS version 1.3.0.4 Index]