AppendPatientLists {OCTOPUS}	R Documentation

AppendPatientLists

Description

This function takes two lists of simulated outcome (eg from SimulateAllPatientOutcomes) and appends such that the structure of the returned list is the same as the source lists. For elements that are a matrix it will utilize rbind, for vectors c() is used. The main purpose of this function is during simulaitons if the original SimulateAllPatients did not generate enought patients

Usage

AppendPatientLists(lList1, lList2, vNamesToExclude = NULL)

Arguments

lList1
- The first elment list where lList2 is appended to.

lList2
- The second list

vNamesToExclude
a vector of the names to include in the appending. Names not included are coppied from lList1 with nothing appended.

[Package OCTOPUS version 1.3.0.4 Index]