CheckGlobalVariables {OCTOPUS}	R Documentation

CheckGlobalVariables

Description

Determine if the global variables that are needed have been defined and if not define them. Use of global variables should be restricted to the variables that are used throughout the code and only in VERY limited use.

Usage

CheckGlobalVariables()

Arguments

lData
- The dataset (as a list) that you want to select from

lValue
- A list with the desired values for each column in lData

[Package OCTOPUS version 1.3.0.4 Index]