ComputePosteriorProbs.MAVOnly {OCTOPUS}	R Documentation

ComputePosteriorProbs.MAVOnly

Description

Used in a trial with a Bayesian analysis and parameter of interest θ, where the decisions are based one posterior probabilities Pr( θ_P - θ_T > MAV |data ). This function returns a list with dPrGrtMAV, dMAVCutoff and dTVCutoff. 0 <= dMAVCutoff and dTVCutoff <= 1

Usage

## S3 method for class 'MAVOnly'
ComputePosteriorProbs(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis, lSamples)

Arguments

nISAAnalysisIndx
A the index of the analysis

bIsFinalISAAnalysis
TRUE or FALSE to indicate if this is the final analysis. Typically, used in a Bayesian design such that the final analysis can have different cutoff values. dTVCutoff

lSamples
List with two vectors, vPostSampPlac and vPostSampTrt, that are samples of the posterior of θ for the placebo and treatment, respectively.

cAnalysis$vMAVCutoff
A vector of cutoffs for the MAV at each analysis.

cAnalysis$vTVCutoff
A vector of cutoffs for the TV.

Value

List with three values dPrGrtMAV, dMAVCutoff and dTVCutoff

[Package OCTOPUS version 1.3.0.4 Index]