ComputePosteriorProbs.default {OCTOPUS}	R Documentation

ComputePosteriorProbs.default

Description

Used in a trial with a Bayesian analysis and parameter of interest θ, where the decisions are based two posterior probabilities Pr( θ > MAV |data ) > vMAVCutoff and Pr( θ > TV | data ) > vTVCutoff. This function returns a list with dMAVCutoff and dTVCutoff. 0 <= dMAVCutoff and dTVCutoff <= 1

Usage

## Default S3 method:
ComputePosteriorProbs(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis, lSamples)

Arguments

nISAAnalysisIndx
A the index of the analysis

bIsFinalISAAnalysis
TRUE or FALSE to indicate if this is the final analysis. Typically, used in a Bayesian design such that the final analysis can have different cutoff values.

lSamples
List with two vectors, vPostSampPlac and vPostSampTrt, that are samples of the posterior of θ for the placebo and treatment, respectively.

cAnalysis$vMAVCutoff
A vector of cutoffs for the MAV at each analysis.

cAnalysis$vTVCutoff
A vector of cutoffs for the TV.

Value

List with two valued dMAVCutoff and dTVCutoff

List with four values dPrGrtMAV, dPrGrtTV, dMAVCutoff and dTVCutoff

[Package OCTOPUS version 1.3.0.4 Index]