CreateDataSetForAnalysis.default {OCTOPUS}	R Documentation

CreateDataSetForAnalysis.default

Description

This function creates a dataset for analysis based on the current time (dCurrentTime) and returns an object with class = "TrialData" that has a list with an element for each ISA data. Before creating the dataset the ProcessData to create the version of the data needed, eg change from baseline, and also coppies covariates that are present.

Usage

## Default S3 method:
CreateDataSetForAnalysis(cPats, dCurrentTime, vISAStatus)

Arguments

cPats
The object of patient data to create the data for analysis from.

dCurrentTime
The current time in the trial. In each simulated trial this will be the time at which the analysis is created.

vISAStatus
A vector with one element for each ISA. If the status of the ISA != 0 then a dataset is created, if the status = 0 for an ISA the dataset will not be created.

Value

An object with class = "TrialData" that has a list with an element for each ISA data. Before creating the dataset the ProcessData to create the version of the data needed, eg change from baseline, and also coppies covariates that are present.

[Package OCTOPUS version 1.3.0.4 Index]