GetBayesianCutoffs.MAVTarget {OCTOPUS}	R Documentation

GetBayesianCutoffs.MAVOnly

Description

This function is the same as GetBayesianCutoffs.MAVOnly as MAVTarget inheirts from MAVOnly

Usage

## S3 method for class 'MAVTarget'
GetBayesianCutoffs(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis)

Arguments

cAnalysis$vPUpper
Vector of upper cutoffs, typically used in the context of success if a posterior probability > vPUpper, 0 <= vPLowerCutoff <= vPUpperCutoff <= 1

cAnalysis$vPLower
Vector of lower cutoffs, typically used in the context of failure if a posterior probability < vPLower, 0 <= vPLowerCutoff <= vPUpperCutoff <= 1

cAnalysis$dFinalPUpper
Value of upper cutoff used at the FINAL analysis, typically used in the context of success if a posterior probability > dFinalPUpper, 0 <= dPLowerCutoff <= dPUpperCutoff <= 1

cAnalysis$dFinalPLower
Value of lower cutoff used at the FINAL analysis, typically used in the context of failure if a posterior probability < dFinalPLower, 0 <= dPLowerCutoff <= dPUpperCutoff <= 1

[Package OCTOPUS version 1.3.0.4 Index]