GetBayesianCutoffs.default {OCTOPUS}	R Documentation

GetBayesianCutoffs.default

Description

Used in a trial with a Bayesian analysis and parameter of interest θ, where the decisions are based two posterior probabilities Pr( θ > MAV |data ) > vMAVCutoff and Pr( θ > TV | data ) > vTVCutoff. This function returns a list with dMAVCutoff and dTVCutoff. 0 <= dMAVCutoff and dTVCutoff <= 1. Typically, this function is not called directly but as part of the ComputePosteriorProb function.

Usage

## Default S3 method:
GetBayesianCutoffs(cAnalysis, nISAAnalysisIndx,
  bIsFinalISAAnalysis)

Arguments

nISAAnalysisIndx
An integer with the index of the analysis

bIsFinalISAAnalysis
TRUE or FALSE to indicate if this is the final analysis. Typically, used in a Bayesian design such that the final analysis can have different cutoff values.

cAnalysis$vMAVCutoff
A vector of cutoffs for the MAV at each analysis.

cAnalysis$vTVCutoff
A vector of cutoffs for the TV.

Value

List with two values dMAVCutoff and dTVCutoff.

[Package OCTOPUS version 1.3.0.4 Index]