GetLSDiffGLS {OCTOPUS}	R Documentation

GetLSDiffGLS

Description

This function will produce the estiamtes comparing to treatment 0 for the LME models that are fit. The Z values will match to SAS, which is what is used in decision making. This version requires glsFit requires glsFit to be a fit for gls

Usage

GetLSDiffGLS(glsFit, nTrt, nTime, dMAV, dTV, dLowerCI, dUpperCI,
  bPlacMinusTrt)

Arguments

glsFit:
The results of the lme() model fit, required to have the variable TRT and Time

nTrt:
Integer value for the Treatment to compare to 0 (placebo)

nTime:
Integer value for the time variable that should be use. For example 24, for 24

[Package OCTOPUS version 1.3.0.4 Index]