InitializeISARandomizer.EqualRandomizer {OCTOPUS}	R Documentation

InitializeISARandomizer.EqualRandomizer

Description

This function is used to initialize the EqualRandomzier for an ISA. It uses the vTrtLab and vQtyPats to sample the desired number of each treatment in the ISA.

Usage

## S3 method for class 'EqualRandomizer'
InitializeISARandomizer(cISARandomizer,
  dISAStartTime)

Arguments

cISARandomizer
Must contain elements cISARandomizer$vTrtLab and cISARandomizer$vQtyPats and the lengths must be equal. If cISARandomizer contains a definition of cISARandomizer$dfSubgroupEnrollmentStatus it is used as the enrollment to a specific covariate signature.

dISAStartTime
Parameter is not utilized for EqualRandomizer

[Package OCTOPUS version 1.3.0.4 Index]