InitializeISARandomizer.POCRandomizer {OCTOPUS}	R Documentation

InitializeISARandomizer.POCRandomizer

Description

InitializeISARandomizer.POCRandomizer will randomize the first patients to a select set of arms or doses in the ISA (eg a POC phase). The remaining arms or doses are opened based on the number of intitial patients.

Usage

## S3 method for class 'POCRandomizer'
InitializeISARandomizer(cISARandomizer,
  dISAStartTime)

Arguments

cISARandomizer
The cISARandomizer$vQtyPatsInit identifies the initial patients to randomize first, the vQtyPatsInit is part of the vQtyPats for that ISA and is not required by other randomizer so this function will stop if it is not defined

dISAStartTime
Parameter is not utilized for POCRandomizer

[Package OCTOPUS version 1.3.0.4 Index]