InitializePatientList {OCTOPUS}	R Documentation

InitializePatientList

Description

This function will create a list to store information about the current patients in the trial. The list will contain the following elements: vStartTimes The start times for each patient in the trial, vTrt Vector of treatment labels in the platform vISA Vector of ISA labels in the trial, lPatOut patient outcomes as a list (see more below), vCurrentQtyPatsISA The number of patients in each ISA, vTrtLab The treatment labeled, taken directly from the cTrialDesign, vISALab The ISA labels, taken directly from the cTrialDesign vQtyPatsArmISA Number of patients in each arm in each ISA

lPatOut A list of patient outcomes. The list will contain a mOutXX structure for each outcome where XX is the outcome number vObsTimeXX The observed time for each outcome. The class( mOutXX ) should be the same as the vISAAnalysis[[ XX ]]

For multiple ISAs, the first ISA is used to determine how to process the outcome data. Specifically, the class elements are copied from the first ISA analysis object vISAAnalysis. That is, the class of mOutXX is class( cTrialDesign$cISADesigns[[1]]$cISAAnalysis$vAnalysis[[ XX ]] )

The returned list will not have any patients in it but will be initialized.

Usage

## Default S3 method:
InitializePatientList(cTrialDesign)

[Package OCTOPUS version 1.3.0.4 Index]