InitializeTrialRandomizer.default {OCTOPUS}	R Documentation

InitializeTrialRandomizer.default

Description

This function is used to initialize the trial randomizer. The default version each ISA and calls the InitializeISARandomizer. Since each ISA can have a specific randomizer the default for the TrialRandomizer is just to initialize each ISA randomizer.

Usage

InitializeTrialRandomizer(cTrialDesign, vISAStartTimes)

Arguments

cTrialDesign
Trial design opbject giving specifics about the trial design.

vISAStartTime
A vector of times that each ISA opens/starts in the trial. A vector of c(0,...) would indicate all ISAs start at the start of the trial.

[Package OCTOPUS version 1.3.0.4 Index]