MakeDecision.TwoOutcomeOption3 {OCTOPUS}	R Documentation

MakeDecision.TwoOutcomeOption3

Description

This two outcome decision options makes if GO if either outcome is a Go and makes a No Go if BOTH outcomes are a No Go, otherwise a Pause is made. For each outcome you make a decision Go (G), Pause (P) or No Go (NG) and when the two outcomes are combined the following tabled describes how decisions will be made.

G | 	 G 	 G 	 G

Outcome 2 	 P | 	 P 	 P 	 G

NG| 	 NG 	 P 	 G

--- 	--- 	 --- 	---

NG 	 P 	 G

Outcome 1

Usage

## S3 method for class 'TwoOutcomeOption3'
MakeDecision(lDecision, lResAnalysis,
  bFinalAnalysis, cRandomizer)

[Package OCTOPUS version 1.3.0.4 Index]