MakeDecisionDoses.AtLeastOne {OCTOPUS}	R Documentation

MakeDecisionDoses.AtLeastOne

Description

This option allows for mulitiple doses and if all doses reach a No Go decision then the combined decision is a No Go. If ANY dose is a Go then a Go decision is made.

Usage

## S3 method for class 'AtLeastOne'
MakeDecisionDoses(lDoseDec)

[Package OCTOPUS version 1.3.0.4 Index]