MakeDecisionDoses.default {OCTOPUS}	R Documentation

MakeDecisionDoses.default

Description

Generic function to make decisions when doses are present in an ISA. No default defined so the simulations stop if a specific approach is not defined.

Usage

## Default S3 method:
MakeDecisionDoses(lDoseDec)

[Package OCTOPUS version 1.3.0.4 Index]