MakeTrialDecision {OCTOPUS}	R Documentation

MakeTrialDecision

Description

This function determines the status of a trial (all ISAs) based on the input. Each ISA can be in one of eight states. The following list provides the details of the returned vISAStatus and what they mean.

Status 0

The ISA is not open, the virtual trial is at a point prior to the ISA opening.

Status 1

The ISA is open for enrollment.

Status 2

The ISA has met the maximum enrollment.

Status 3

The ISA was closed with a Go BEFORE Final Analysis (FA). This can be a Final status for an ISA because it was closed.

Status 4

The ISA was closed with a No Go BEFORE FA. This can be a Final status for an ISA because it was closed.

Status 5

The ISA was closed with a Go at FA. This can be a Final status for an ISA because it was closed.

Status 6

The ISA was closed with a No Go at FA. This can be a Final status for an ISA because it was closed.

Status 7

The ISA was closed with a Pause at FA. This can be a Final status for an ISA because it was closed.

The Status options of 3 or 4 indicate the ISA was closed early, before the FA, either for a Go or a No Go. The options 5-7 all were closed at the FA.

Usage

## Default S3 method:
MakeTrialDecision(cISADesigns, lResAnalysis, vISAStatus,
  vIsFinalISAAnalysis, cRandomizer)

[Package OCTOPUS version 1.3.0.4 Index]