OCTOPUS {OCTOPUS}	R Documentation

OCTOPUS

Description

The OCTOPUS package is used to simulate platform clinical trials. Platform trials are clinical trials that are designed with the flexibility to add new interventions as the trial progresses.

License

This package is governed by the JRD OCTOPUS License, which is the GNU General Public License V3 with additional terms. The precise license terms are located in the files LICENSE and GPL.

To view license you may use the following command:

file.show( system.file("LICENSE", package="OCTOPUS"))

file.show( system.file("GPL", package="OCTOPUS"))

Introduction

General intro to the packages ..

Basic description

Stuff

Author(s)

Maintainer: J. Kyle wathen kylewathen@gmail.com

See Also

Useful links:

https://github.com/kwathen/OCTOPUS

[Package OCTOPUS version 1.3.0.4 Index]