ProcessData {OCTOPUS}	R Documentation

ProcessData

Description

This function is used to process the data before the RunAnalysis function is called. This can be useful for cases where the data is simulated/collected in one form, but another form is needed for analysis. For example, the patient data over time is simulated but the data is analysed as change from baseline.

Usage

ProcessData(cDS, dCurrentTime)

[Package OCTOPUS version 1.3.0.4 Index]