ProcessData.ProcessReptMeasChngBaseline {OCTOPUS}	R Documentation

ProcessData.ProcessReptMeasChngBaseline

Description

This function is used to process the data before the RunAnalysis function is called. This can be useful for cases where the data is simulated/collected in one form but another form is needed for analysis. For example, the patient data over time is simulated but the data is analysed as change from baseline. This function will use the vObsTime and vStartTime to remove patients that have not enrolled and set outcomes to NA that have not been observed Will return the change from baseline dCurrentTime, vTrt, mPatOut, lSimOut, vStartTimes

Usage

## S3 method for class 'ProcessReptMeasChngBaseline'
ProcessData(cDS, dCurrentTime)

[Package OCTOPUS version 1.3.0.4 Index]