Randomize {OCTOPUS}	R Documentation

Randomize

Description

Randomize should return a list with at least the following items nTrt - The treatment within the ISA that the patient should receive nISA - The index of the ISA cISARand - the same is cISARand but if it is updates this allows the caller to get the update. For a platform trial the randomization is a two step process. The two steps are as follows: 1. First randomize between the ISAs. To do this RandomizeBetweenISA is called and the class( cRandomizer ) will determine which version of this function is called. For example, if class( cRandomzier ) = "Equal" see RandomizeBetweenISA.Equal() for that definition. Each ISA RandomzieBetweenISA function will return an integer nISA to indicate which ISA the patient was randomized to. 2. Randomize within the ISA. To do this RandomizeWithinISA will be called and the class( cRandomzier[[ nISA ]] ) which specific randomizer is called.

Usage

Randomize(cRandomizer, vISAStatus, dCurrentTime, dfCov, nPrintDetail)

[Package OCTOPUS version 1.3.0.4 Index]