ReturnNextTreatment {OCTOPUS}	R Documentation

ReturnNextTreatment

Description

This function uses the first element of cISARand and the first item in the vector and as the assigned treatment. The cISARand in the return will have this element removed.

Usage

ReturnNextTreatment(cISARand)

[Package OCTOPUS version 1.3.0.4 Index]