RunExample {OCTOPUS}	R Documentation

RunExample

Description

This function is used to run the example in the package. For a current list of example call OCTOPUS::RunExample() and you will get a list of the available examples.

There are example(s) that require shiny so you must install the shiny package to run some of the examples.

Usage

RunExample(example)

Examples

## Not run: {OCTOPUS::RunExample( "CompareRecruitment" )
}
## End(Not run)

[Package OCTOPUS version 1.3.0.4 Index]