RunTrialAnalysis {OCTOPUS}	R Documentation

RunTrialAnalysis

Description

RunTrialAnalysis is a generic method to execute the trial level analysis, eg execute all ISA analysis. Any implementation of RunTrialAnalysis must increment vISAAnalysisIndx if an analysis is run for an ISA

RunTrialAnalysis is a generic method to execute the trial analysis. Any implementation of RunTrialAnalysis must increment vISAAnalysisIndx if an analysis is run for an ISA. Trial level analysis - will go through and conduct the analysis for each ISA. This version is sent all the available data in lDataAna then for each ISA gets the subset that is specific to that ISA. If we want to do a version that borrows control info a new version could be adapted. Because the decision is tied to the analysis, each ISA analysis should contain the elementsnGo, nNoGo, nPause as part ot the list retured.

Usage

RunTrialAnalysis(cISADesigns, cEnrolledPats, vISAStatus, dCurrentTime,
  vRunISAAnalysis, vISAAnalysisIndx, vIsFinalISAAnalysis, cRandomizer)

## Default S3 method:
RunTrialAnalysis(cISADesigns, cEnrolledPats, vISAStatus,
  dCurrentTime, vRunISAAnalysis, vISAAnalysisIndx, vIsFinalISAAnalysis,
  cRandomizer)

[Package OCTOPUS version 1.3.0.4 Index]