SimPatientOutcomes {OCTOPUS}	R Documentation

SimPatientOutcomes

Description

SimPatientOutcomes This function is intended to simulate the outcomes for a given ISA.

Usage

SimPatientOutcomes(cSimOutcomes, cISADesign, dfPatCovISA)

## Default S3 method:
SimPatientOutcomes(cSimOutcomes, cISADesign, dfPatCovISA)

## S3 method for class 'Independent'
SimPatientOutcomes(cSimOutcomes, cISADesign,
  dfPatCovISA)

## S3 method for class 'Correlated'
SimPatientOutcomes(cSimOutcomes, cISADesign,
  dfPatCovISA)

Methods (by class)

default: This function is intended to simulate the outcomes for a given ISA. Because several options are provided and there in no well defined default an stop error occurs if you call the default method.

Independent: The cSimOutcomes object has a list of sim outcomes and this version will go through and call the SimPatientOutcomes for each type. Each outcome is simulated independently.

Correlated: The cSimOutcomes object has a list of sim outcomes and this version will utilize the first outcome to simulate all outcomes correlated. Each simoutcome must have an attribute vColIndex that specifies which columns out of the matrix will be used for that outcome.

[Package OCTOPUS version 1.3.0.4 Index]