SimulateAllPatientCovariates {OCTOPUS}	R Documentation

SimulateAllPatientCovariates

Description

SimulateAllPatientCovariates This function is used to simulate all patient covariates in the platform.

SimulateAllPatientCovariates This function is used to simulate all patient covariates in the platform. This function is generic in case there is a need to override but is implemented such that it should be suitable for most cases. Specifically, this function calls SimulatePatientCovariates for each ISA.

Usage

SimulateAllPatientCovariates(cSimCovariates, cTrialDesign)

## Default S3 method:
SimulateAllPatientCovariates(cSimCovariates,
  cTrialDesign)

Value

A list with length equal to the number of ISAs in the trial. Each element represents the simulated covariates for each ISA.

[Package OCTOPUS version 1.3.0.4 Index]