SimulateAllPatientOutcomes {OCTOPUS}	R Documentation

SimulateAllPatientOutcomes

Description

SimulateAllPatientOutcomes This function is used to simulate all patient outcomes in the platform.

Usage

SimulateAllPatientOutcomes(cScen, cTrialDesign, dfPatCov)

## Default S3 method:
SimulateAllPatientOutcomes(cScen, cTrialDesign, dfPatCov)

Methods (by class)

default: This function is used to simulate all patient outcomes in the platform. This function is generic in case there is a need to override but is implemented such that it should be suitable for most cases. Specifically, this function calls SimPatientOutcomes for each ISA.

[Package OCTOPUS version 1.3.0.4 Index]