SimulatePatientCovariates.Categorical {OCTOPUS}	R Documentation

SimulatePatientCovariates.Categorical

Description

SimulatePatientCovariates This function is used to simulate the patient covariate values. This function is intended to do one covariate for all patients.

Usage

## S3 method for class 'Categorical'
SimulatePatientCovariates(cSimCovariate,
  nQtyOfPatients)

SimulatePatientCovariates(cSimCovariate, nQtyOfPatients)

## Default S3 method:
SimulatePatientCovariates(cSimCovariate, nQtyOfPatients)

Methods (by class)

Categorical: This function is used to simulate the patient covariate values. This function is intended to do one covariate, for all patients. The sampled covaraites values will be 1,2,..., length( cSimCovariate$vProbs ).

default: This function is used to simulate the patient covariate values. This function is intended to do one covariate, for all patients. Because several options are provided and there in no well defined default an stop error occurs if you call the default method.

[Package OCTOPUS version 1.3.0.4 Index]