ggPlotPlatformRecruitment {OCTOPUS}	R Documentation

PlotPlatformRecruitment

Description

This function creates a plot of recruitment comparing a platform with 2 ISAs and 2 consecutive POC studies. Mainly used by the Shiny app execute by calling RunExample( "CompareRecruitment" ). This version returns a ggplot2 object.

Usage

ggPlotPlatformRecruitment(nQtyReps, nMaxQtyPats, vPatsPerMonthPerSite1,
  dDelayToStartPOC, dDelayBetweenTrialsPOC, vQtySitesPOC,
  dDelayToStartPlat, dDelayBetweenTrialsPlat, vQtyOfSitesPlat1,
  vQtyOfSitesPlat2)

[Package OCTOPUS version 1.3.0.4 Index]